+++ src/app/store/task_store.js
@@ -2,6 +2,7 @@
 
 const Suite = require('../common/suite');
 const TasksHandler = require('../api/tasks_handler');
+const EnvVariableHandler = require('../api/env_variables_handler');
 const TaskImporter = require('../common/task_importer');
 
 const AppSettings = require('../../app_settings.json');
@@ -10,7 +11,8 @@ module.exports = {
     state: {
         suites: [],
         selectedSuite: 0,
-        runningTasks: 0
+        runningTasks: 0,
+        globalEnv: []
     },
     getters: {
         suites(state) {
@@ -89,6 +91,9 @@ module.exports = {
                 task.title = suite.getValidTaskName(task.title);
             }
         },
+        setGlobalEnv(state, data) {
+            state.globalEnv = data;
+        },
         _setSuites(state, suites) {
             state.suites.splice(0, state.suites.length);
             suites.forEach((suite) => {
@@ -110,9 +115,15 @@ module.exports = {
         saveTasks(context) {
             TasksHandler.saveTasks(context.getters.suites);
         },
+        saveGlobalEnvVariables(context) {
+            const variables = EnvVariableHandler.saveGlobalEnv(context.state.globalEnv);
+            context.commit("setGlobalEnv", variables);
+        },
         loadTasks(context) {
             const loadedSuites = TasksHandler.loadTasksFromConfig();
             context.commit("_setSuites", loadedSuites);
+            const loadedGlobalEnv = EnvVariableHandler.loadGlobalEnvFromConfig();
+            context.commit("setGlobalEnv", loadedGlobalEnv);
         },
         stopAllTasks(context) {
             const promises = context.getters.suites.map((s) => s.stopAll());