+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Overhaul of menu and context menu options
     * Suite-related actions moved to context menu
     * "Import suite" is not part of settings
+* Global Env variables in settings
 * Adds cut-copy-paste items to context menu
 * Settings shortcuts menu is now collapsed by default
 * Minor fix in "Desktopify"